requirements.txt
asgiref==3.6.0
Django==4.2.1
django-taggit==3.0.0
Markdown==3.4.3
psycopg2-binary==2.9.3
sqlparse==0.4.4
